feat(layout): center content and add scrolling animation
Centers the content for the About and Blog sections to improve visual balance.

Introduces a horizontal scrolling marquee animation for the badges in the About section to create a dynamic element.

Removes the `min-h-screen` class from the main sections to allow for more flexible and content-driven section heights. This also addresses issues with the layout on landscape mode for mobile devices, ensuring the content fits the screen without excessive zooming.